YT-28280: fix integer overflow commit_hash:3a3b72f3a173b86b89b08ca2dbcc1ec2fda7bfe4

+++ yt/yt/server/node/exec_node/job_fs_secretary.cpp
@@ -107,8 +107,8 @@ TError CheckDockerImage(
 }
 
 TError CheckRootVolumeDiskSpaceAndInodeLimit(
-    std::optional<int> baselineDiskSpace,
-    std::optional<int> currentDiskSpace,
+    std::optional<i64> baselineDiskSpace,
+    std::optional<i64> currentDiskSpace,
     std::optional<i64> baselineInodeLimit,
     std::optional<i64> currentInodeLimit)
 {
@@ -771,12 +771,12 @@ void TJobFSSecretary::SetVirtualSandboxReader(NNbd::IImageReaderPtr reader)
     VirtualSandboxData_->Reader = std::move(reader);
 }
 
-const std::optional<int>& TJobFSSecretary::GetRootVolumeDiskSpace() const
+const std::optional<i64>& TJobFSSecretary::GetRootVolumeDiskSpace() const
 {
     return RootVolumeDiskSpace_;
 }
 
-const std::optional<int64_t>& TJobFSSecretary::GetRootVolumeInodeLimit() const
+const std::optional<i64>& TJobFSSecretary::GetRootVolumeInodeLimit() const
 {
     return RootVolumeInodeLimit_;
 }
